refactor: create EmailLoginForm for UserLogin
